moved styles into components folder
