+++ package.json
@@ -31,7 +31,7 @@
   "devDependencies": {
     "browserify": "^13.0.0",
     "watchify": "^3.7.0",
-    "css-modulesify": "^0.17.0",
+    "css-modulesify": "^0.21.0",
     "flux": "^2.1.1",
     "immutable": "^3.7.4",
     "markdown-chalk": "^1.4.0",
